Phase 6d: carve display mode out of app.js into display.ts
display.ts owns the audience view: the `display.active` flag (shared
state object read by app.js's ~14 gating sites), detectDisplayMode
(returns the URL session id for app.js to adopt), the 3s polling loop
with the round-transition overlay + staggered fade-in sequence,
applyDisplayModeUI, initDisplayMode(sessionId), and the display-only
tiebreak panel (renderDisplayModeTiebreak, typed against
GameStateResponse['tiebreak']).

The render pipeline and queue animation stay in app.js and are injected
via initDisplayDeps(): fetchState, renderFromState (plus a
skip-queue-animation variant), hasPendingQueueAnimation /
triggerPendingQueueAnimation, endCompetition, showBattleScreen.
roundTransitionInProgress became module-private (it had no readers
outside the display cluster).

Verified: build + tsc + eslint clean; 9-step Playwright smoke passed;
two-context browser check (admin drives a battle, display context at
/battle/<id>?mode=display) confirming display-mode detection, hidden
voting/controls, and the poll-driven auto-update to round 2 through the
transition sequence — zero JS errors.
